Updated quantity string format and added missing Recipe activity implementation classes.

+++ src/com/aaron/recipe/bean/Fraction.java
@@ -77,7 +77,7 @@ public class Fraction implements Serializable
             }
         }
 
-        return String.valueOf(number);
+        return String.valueOf((int)number); // Cast to int to remove decimal point
     }
 
     /**
